+++ NEWS.md
@@ -9,6 +9,7 @@
 * `any_table()`: Added `row_pct` and `col_pct` keywords to `pct_group` parameter. With this the function can calculate total percentages for rows and columns regardless of the respective other dimension. (21.02.2026)
 * `rename_multi()`: Variable names can now be passed without quotation marks. (28.02.2026)
 * `any_table()`: Percentages based on single formatted variable expressions can now be computed with the `pct_value` parameter. (05.03.2026)
+* `any_table()`: If the `pct_value or `pct_group` parameters are used but are not part of the statistics parameter, they will be automatically added to statistics. (05.03.2026)
 
 ### Changed functionality